Move EtsyOAuthClient from etsy/_v2.py to etsy/oauth.py. This way clients that are read-only can opt to not import etsy.oauth and not need oauth2 installed.
