+++ etsy/__init__.py
@@ -1,5 +1,5 @@
 from _v1 import EtsyV1, EtsyV1
-from _v2 import EtsyV2 as Etsy, EtsyOAuthClient
+from _v2 import EtsyV2 as Etsy
 
 
 __version__ = '0.2.1'
